tests/submission/Infraware/Forms/contents/Forms/input_dirname.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_dirname</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="php/dirname.php"
		  id="input_form">
	<p><input type='hidden' id='input_text' name='input_text'></p>
	</form>

	<script>
		var input_text = document.getElementById("input_text");
		input_text.dirname = 'input_dirname';

		if (typeof(input_text.dirname) == "string") {
			test(function() {
				assert_equals(input_text.dirname, "input_dirname", "dirname attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("dirname attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
